Add util.js and refactor actions
separate action files for index.js
