+++ src/actions/common.js
@@ -1,0 +1,6 @@
+import * as types from '../constants/ActionTypes'
+
+export const changeSaveStatus = status => ({
+  type: types.CHANGE_SAVE_STATUS,
+  status
+})
