+++ src/lib/util.js
@@ -1,0 +1,7 @@
+export const delay = (timeout, cb)  => {
+  setTimeout(() => cb(), timeout)
+}
+
+export const capitalizeFirstLetter = str => (
+  str.charAt(0).toUpperCase() + str.slice(1)
+)
